% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{answer_hw1_2c}
\alias{answer_hw1_2c}
\title{submit answer to homework 1 question 2 part c}
\usage{
answer_hw1_2c(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 2 part c
}
\description{
submit answer to homework 1 question 2 part c
}
\examples{
\donttest{

# format answer to question 2 part c like:
answer_hw1_2c(answer = .2239)

answer_hw1_2c(answer = .5483)

}
}
