% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{answer_hw1_4b}
\alias{answer_hw1_4b}
\title{submit answer to homework 1 question 4 part b}
\usage{
answer_hw1_4b(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 4 part b
}
\description{
submit answer to homework 1 question 4 part b
}
\examples{
\donttest{

# format answer to question 4 part b like:
answer_hw1_4b(answer = TRUE) # to indicate true

answer_hw1_4b(answer = FALSE) # to indicate false


}
}
