% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{answer_hw2_2}
\alias{answer_hw2_2}
\title{submit answer to homework 2 question 2}
\usage{
answer_hw2_2(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 2
}
\description{
submit answer to homework 2 question 2
}
\examples{
\donttest{

# format answer to question 2:
answer_hw2_2(answer = TRUE) # to indicate true

answer_hw2_2(answer = FALSE) # to indicate false


}
}
