% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{answer_hw2_3}
\alias{answer_hw2_3}
\title{submit answer to homework 2 question 3}
\usage{
answer_hw2_3(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 3
}
\description{
submit answer to homework 2 question 3
}
\examples{
\donttest{

# format answer to question 3:
answer_hw2_3(answer = TRUE) # to indicate true

answer_hw2_3(answer = FALSE) # to indicate false


}
}
