% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw2.R
\name{answer_hw2_5}
\alias{answer_hw2_5}
\title{submit answer to homework 2 question 5}
\usage{
answer_hw2_5(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 5
}
\description{
submit answer to homework 2 question 5
}
\examples{
\donttest{

# format answer to question 5 like:
answer_hw2_5(answer = TRUE) # to indicate true

answer_hw2_5(answer = FALSE) # to indicate false


}
}
