% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{answer_hw3_1b}
\alias{answer_hw3_1b}
\title{submit answer to homework 3 question 1 part b}
\usage{
answer_hw3_1b(model0 = NULL, model1 = NULL, model2 = NULL, model3 = NULL)
}
\arguments{
\item{model0}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}

\item{model1}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}

\item{model2}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 1 part b
}
\description{
submit answer to homework 3 question 1 part b
}
\examples{
\donttest{

# format answer to question 1 part b like:
answer_hw3_1b(model0 = .2239, model1 = .05, model2 = .8483)


}

}
