% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{answer_hw3_2b}
\alias{answer_hw3_2b}
\title{submit answer to homework 3 question 2 b}
\usage{
answer_hw3_2b(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 2 b
}
\description{
submit answer to homework 3 question 2 b
}
\examples{
\donttest{

# format answer to question 2:
answer_hw3_2b(answer = TRUE) # to indicate true

answer_hw3_2b(answer = FALSE) # to indicate false


}
}
