% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw3.R
\name{answer_hw3_3}
\alias{answer_hw3_3}
\title{submit answer to homework 3 question 3}
\usage{
answer_hw3_3(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 3
}
\description{
submit answer to homework 3 question 3
}
\examples{
\donttest{

# format answer to question 3:
answer_hw3_3(answer = .2239)

answer_hw3_3(answer = .5483)

}

}
