% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4.R
\name{answer_hw4_2b}
\alias{answer_hw4_2b}
\title{submit answer to homework 4 question 2 part b}
\usage{
answer_hw4_2b(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 2b
}
\description{
submit answer to homework 4 question 2 part b
}
\examples{
\donttest{

# format answer to question 2 part b like:
answer_hw4_2b(answer = .2239)

answer_hw4_2b(answer = .5483)

}

}
