% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4.R
\name{answer_hw4_4}
\alias{answer_hw4_4}
\title{submit answer to homework 4 question 4}
\usage{
answer_hw4_4(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 4
}
\description{
submit answer to homework 4 question 4
}
\examples{
\donttest{

# format answer to question 4:
answer_hw4_4(answer = .2239)

answer_hw4_4(answer = .5483)

}

}
