% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{answer_hw5_1b}
\alias{answer_hw5_1b}
\title{submit answer to homework 5 question 1part b}
\usage{
answer_hw5_1b(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 5b
}
\description{
submit answer to homework 5 question 1part b
}
\examples{
\donttest{

# format answer to question 4:
answer_hw5_1b(answer = .2239)

answer_hw5_1b(answer = .5483)

}

}
