% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{answer_hw5_3}
\alias{answer_hw5_3}
\title{submit answer to homework 5 question 3}
\usage{
answer_hw5_3(answer = NULL)

answer_hw5_3(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 3

summary of your current answer for question 3
}
\description{
submit answer to homework 5 question 3

submit answer to homework 5 question 3
}
\examples{
\donttest{

# format answer to question 5:
answer_hw5_3(answer = .2239)

answer_hw5_3(answer = .5483)

}

\donttest{

# format answer to question 5:
answer_hw5_3(answer = .2239)

answer_hw5_3(answer = .5483)

}

}
