% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{answer_hw5_6}
\alias{answer_hw5_6}
\title{submit answer to homework 5 question 6}
\usage{
answer_hw5_6(answer = NULL)
}
\arguments{
\item{answer}{a vector of length 19}
}
\value{
summary of your current answer for question 6
}
\description{
submit answer to homework 5 question 6
}
\examples{
\donttest{

# format answer to question 6:

grid <- c(1, 2, 3, 4, 5) # should be length 19!
answer_hw5_5(answer = grid)


}

}
