% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{answer_hw5_7}
\alias{answer_hw5_7}
\title{submit answer to homework 5 question 7}
\usage{
answer_hw5_7(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 7
}
\description{
submit answer to homework 5 question 7
}
\examples{
\donttest{

# format answer to question 2:
answer_hw5_7(answer = TRUE) # to indicate true

answer_hw5_7(answer = FALSE) # to indicate false


}
}
