% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw5.R
\name{answer_hw5_7a}
\alias{answer_hw5_7a}
\title{submit answer to homework 5 question 7a}
\usage{
answer_hw5_7a(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 7a
}
\description{
submit answer to homework 5 question 7a
}
\examples{
\donttest{

# format answer to question 7a:
answer_hw5_7a(answer = TRUE) # to indicate true

answer_hw5_7a(answer = FALSE) # to indicate false


}
}
