% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_2}
\alias{answer_hw6_2}
\title{submit answer to homework 6 question 2}
\usage{
answer_hw6_2(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 2
}
\description{
submit answer to homework 6 question 2
}
\examples{
\donttest{

# format answer to question 2:
answer_hw6_2(answer = .2239)

answer_hw6_2(answer = .5483)

}

}
