% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_3a}
\alias{answer_hw6_3a}
\title{submit answer to homework 6 question 3 part a}
\usage{
answer_hw6_3a(answer = NULL)
}
\arguments{
\item{answer}{a vector}
}
\value{
summary of your current answer for question
}
\description{
submit answer to homework 6 question 3 part a
}
\examples{
\donttest{

# format answer to question 3a:

grid <- c(1, 2, 3, 4, 5) # should be length 19!
answer_hw6_3a(answer = grid)


}

}
