% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_4}
\alias{answer_hw6_4}
\title{submit answer to homework 6 question 4}
\usage{
answer_hw6_4(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 4
}
\description{
submit answer to homework 6 question 4
}
\examples{
\donttest{

# format answer to question 4:
answer_hw6_4(answer = TRUE) # to indicate true

answer_hw6_4(answer = FALSE) # to indicate false


}
}
