% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_5a}
\alias{answer_hw6_5a}
\title{submit answer to homework 6 question 5a}
\usage{
answer_hw6_5a(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 5a
}
\description{
submit answer to homework 6 question 5a
}
\examples{
\donttest{

# format answer to question 5a:
answer_hw6_5a(answer = .2239)

answer_hw6_5a(answer = .5483)

}

}
