% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_5b}
\alias{answer_hw6_5b}
\title{submit answer to homework 6 question 5b}
\usage{
answer_hw6_5b(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 5b
}
\description{
submit answer to homework 6 question 5b
}
\examples{
\donttest{

# format answer to question 5b:
answer_hw6_5b(answer = TRUE) # to indicate true

answer_hw6_5b(answer = FALSE) # to indicate false


}
}
