% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_5c}
\alias{answer_hw6_5c}
\title{submit answer to homework 6 question 5c}
\usage{
answer_hw6_5c(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 5c
}
\description{
submit answer to homework 6 question 5c
}
\examples{
\donttest{

# format answer to question 5c:
answer_hw6_5c(answer = TRUE) # to indicate true

answer_hw6_5c(answer = FALSE) # to indicate false


}
}
