% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{answer_hw6_6}
\alias{answer_hw6_6}
\title{submit answer to homework 6 question 6}
\usage{
answer_hw6_6(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 6
}
\description{
submit answer to homework 6 question 6
}
\examples{
\donttest{

# format answer to question 6:
answer_hw6_6(answer = TRUE) # to indicate true

answer_hw6_6(answer = FALSE) # to indicate false


}
}
