% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_1a}
\alias{answer_proj_1a}
\title{submit answer to project question 1a}
\usage{
answer_proj_1a(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 1a
}
\description{
submit answer to project question 1a
}
\examples{
\donttest{

# format answer to question 1a:
answer_proj_1a(answer = TRUE) # to indicate true

answer_proj_1a(answer = FALSE) # to indicate false


}
}
