% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_1b}
\alias{answer_proj_1b}
\title{submit answer to project question 1b}
\usage{
answer_proj_1b(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 1b
}
\description{
submit answer to project question 1b
}
\examples{
\donttest{

# format answer to question 1b:
answer_proj_1a(answer = TRUE) # to indicate true

answer_proj_1a(answer = FALSE) # to indicate false

}

}
