% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_1d}
\alias{answer_proj_1d}
\title{submit answer to project question 1d}
\usage{
answer_proj_1d(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer, an integer from (1,2,3,4,5)}
}
\value{
summary of your current answer for question 1d
}
\description{
submit answer to project question 1d
}
\examples{
\donttest{

# format answer to question 1d:
answer_proj_1d(answer = 1)

answer_proj_1d(answer = 2)

}

}
