% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_2a}
\alias{answer_proj_2a}
\title{submit answer to project question 2a}
\usage{
answer_proj_2a(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer, an integer from (1,2,3)}
}
\value{
summary of your current answer for question 2a
}
\description{
submit answer to project question 2a
}
\examples{
\donttest{

# format answer to question 2a:
answer_proj_2a(answer = 1)

answer_proj_2a(answer = 2)

}

}
