% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_4c}
\alias{answer_proj_4c}
\title{submit answer to project question 4c}
\usage{
answer_proj_4c(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer, 1, 2 or 3.}
}
\value{
summary of your current answer for question 4c
}
\description{
submit answer to project question 4c
}
\examples{
\donttest{

# format answer to question 4c:
answer_proj_4c(answer = 1)

answer_proj_4c(answer = 2)

}

}
