% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_4d}
\alias{answer_proj_4d}
\title{submit answer to project question 4 part d}
\usage{
answer_proj_4d(answer = NULL)
}
\arguments{
\item{answer}{a vector}
}
\value{
summary of your current answer for question
}
\description{
submit answer to project question 4 part d
}
\examples{
\donttest{

# format answer to question 4d:

grid <- c(1.1, 2.2, 3.3, 4.4, 5.5) # should be length 5!
answer_proj_4d(answer = grid)


}

}
