% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_4e}
\alias{answer_proj_4e}
\title{submit answer to project question 4e}
\usage{
answer_proj_4e(answer = NULL)
}
\arguments{
\item{answer}{a vector of length 5}
}
\value{
summary of your current answer for question 4e
}
\description{
submit answer to project question 4e
}
\examples{
\donttest{

# format answer to question 4e:

grid <- c(1.1, 2.2, 3.3, 4.4, 5.5) # should be length 5!
answer_proj_4d(answer = grid)

}

}
