% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{generate_data_hw1}
\alias{generate_data_hw1}
\title{Load data for homework 1}
\usage{
generate_data_hw1(netid)
}
\arguments{
\item{netid}{a character string representing your NYU netid. Make sure to only use your netID!!}
}
\value{
the `schools` data.frame
}
\description{
Load data for homework 1
}
\examples{
\donttest{

# run code to load your unique dataset
generate_data_hw1(netid = 'gp77') # put net id in quotes


}

}
