% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4.R
\name{generate_data_hw4}
\alias{generate_data_hw4}
\title{Load data for homework 4}
\usage{
generate_data_hw4(netid)
}
\arguments{
\item{netid}{a character string representing your NYU netid. Make sure to only use your netID!!}
}
\value{
the `df` data.frame
}
\description{
Load data for homework 4
}
\examples{
\donttest{

# run code to load your unique dataset
generate_data_hw4(netid = 'gp77') # put net id in quotes


}

}
