% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw6.R
\name{generate_data_hw6}
\alias{generate_data_hw6}
\title{Load data for homework 2}
\usage{
generate_data_hw6(netid)
}
\arguments{
\item{netid}{a character string representing your NYU netid. Make sure to only use your netID!!}
}
\value{
the `schools` data.frame
}
\description{
Load data for homework 2
}
\examples{
\donttest{

# run code to load your unique dataset
generate_data_hw5(netid = 'gp77') # put net id in quotes


}

}
