% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{generate_data_proj}
\alias{generate_data_proj}
\title{Load data for project}
\usage{
generate_data_proj(netid)
}
\arguments{
\item{netid}{a character string representing your NYU netid. Make sure to only use your netID!!}
}
\value{
the `housing` data.frame
}
\description{
Load data for project
}
\examples{
\donttest{

# run code to load your unique dataset
generate_data_proj(netid = 'gp77') # put net id in quotes


}

}
