% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw0.R
\docType{data}
\name{nEnrolled}
\alias{nEnrolled}
\title{Load names for all generate data functions}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
nEnrolled
}
\description{
Load names for all generate data functions
}
\keyword{datasets}
