% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{question_hw1_3a}
\alias{question_hw1_3a}
\title{Reveal question 3a}
\usage{
question_hw1_3a()
}
\value{
the question for 3 part a
}
\description{
Reveal question 3a
}
\examples{
\donttest{

# run code to print the question
question_hw1_3a()


}

}
