% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw1.R
\name{question_hw1_4a}
\alias{question_hw1_4a}
\title{Reveal question 4a}
\usage{
question_hw1_4a()
}
\value{
the question for 4 part a
}
\description{
Reveal question 4a
}
\examples{
\donttest{

# run code to print the question
question_hw1_4a()


}

}
