% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw4.R
\name{question_hw4_4}
\alias{question_hw4_4}
\title{Reveal hw 4 question4}
\usage{
question_hw4_4()
}
\value{
Question 4
}
\description{
Reveal hw 4 question4
}
\examples{
\donttest{

# Run the function

answer_hw4_4()

}

}
