% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_1c}
\alias{answer_proj_1c}
\title{submit answer to project question 1c}
\usage{
answer_proj_1c(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer. Do not worry about rounding but provide a minimum or 2 decimal places to all answers}
}
\value{
summary of your current answer for question 1c
}
\description{
submit answer to project question 1c
}
\examples{
\donttest{

# format answer to question 1d:
answer_proj_1c(answer = .2239)

answer_proj_1c(answer = .5483)

}

}
