% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_2e}
\alias{answer_proj_2e}
\title{submit answer to project question 2e}
\usage{
answer_proj_2e(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 2e
}
\description{
submit answer to project question 2e
}
\examples{
\donttest{

# format answer to question 2e:
answer_proj_2e(answer = TRUE) # to indicate true

answer_proj_2e(answer = FALSE) # to indicate false

}

}
