% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_3a}
\alias{answer_proj_3a}
\title{submit answer to project question 3a}
\usage{
answer_proj_3a(answer = NULL)
}
\arguments{
\item{answer}{a numeric answer, an intergerfrom (1,2,3)}
}
\value{
summary of your current answer for question 3a
}
\description{
submit answer to project question 3a
}
\examples{
\donttest{

# format answer to question 3a:
answer_proj_3a(answer = 1)

answer_proj_3a(answer = 2)

}

}
