% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{answer_proj_4b}
\alias{answer_proj_4b}
\title{submit answer to project question 4b}
\usage{
answer_proj_4b(answer = NULL)
}
\arguments{
\item{answer}{a logical indicating TRUE or FALSE}
}
\value{
summary of your current answer for question 4b
}
\description{
submit answer to project question 4b
}
\examples{
\donttest{

# format answer to question 4b:
answer_proj_4b(answer = TRUE) # to indicate true

answer_proj_4b(answer = FALSE) # to indicate false

}

}
